SettableClock: Refactored setClockTime to immediately execute/skip overdue clock events in clock time order.
Previously, they were scheduled for the same simulation time, but it could terminate the simulation because
the clock events and the corresponding clock/oscillator was left in inconsistent state. The clock was already
set to the new clock time even though the overdue clock events were not yet executed.
When the events were executed they set the clock time back.
This also fixes #844.